Create rmgpy.rmg.listener submodule and move SimulationProfileWriter to this submodule
Didn't seem appropriate for it to be placed in rmgpy.tools.sensitivity
